platform specific changes of osx pointers representation

+++ analyzer/lib/mcs/cpp/object_id.h
@@ -1,9 +1,10 @@
 #ifndef OBJECT_ID_H
 #define OBJECT_ID_H
 
+#include <cstdint>
 #include "hanser_recursive.h"
 
-typedef unsigned int ObjectID;
+typedef uint64_t ObjectID;
 typedef HanserRecursive<ObjectID> HR;
 
 #endif // OBJECT_ID_H
